requirements.txt
blessed==1.19.1
brotlipy==0.7.0
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
contourpy==1.0.6
cryptography @ file:///croot/cryptography_1665612644927/work
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.8.0
fonttools==4.38.0
gitdb==4.0.9
GitPython==3.1.29
gpustat==1.0.0
h5py @ file:///tmp/build/80754af9/h5py_1637138879700/work
huggingface-hub==0.11.0
idna @ file:///croot/idna_1666125576474/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib @ file:///croot/joblib_1666298844297/work
kiwisolver==1.4.4
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.6.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
nvidia-ml-py==11.495.46
packaging==21.3
pandas==1.5.1
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==4.21.9
psutil==5.9.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642617107864/work
scipy @ file:///tmp/build/80754af9/scipy_1618855647378/work
seaborn==0.12.1
sentry-sdk==1.11.0
setproctitle==1.3.2
shortuuid==1.0.11
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tokenizers==0.13.2
torch==1.11.0
torch-cluster @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-cluster_1656603558144/work
torch-geometric @ file:///usr/share/miniconda/envs/test/conda-bld/pyg_1665661116090/work
torch-scatter @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-scatter_1656600927268/work
torch-sparse @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-sparse_1665388423654/work
torchvision==0.12.0
tqdm @ file:///opt/conda/conda-bld/tqdm_1664392687731/work
transformers==4.24.0
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///croot/urllib3_1666298941550/work
wandb==0.13.5
wcwidth==0.2.5
